--- a/backend/src/question/__init__.py
+++ b/backend/src/question/__init__.py
@@ -21,8 +21,15 @@ from .schemas import (
 
 # Services
 from .services import (
-    ContentProcessingService,
+    # Generation service
     QuestionGenerationService,
+    # Content processing functions
+    get_content_from_quiz,
+    get_content_statistics,
+    prepare_and_validate_content,
+    prepare_content_for_generation,
+    validate_content_quality,
+    validate_module_content,
 )
 
 __all__ = [
@@ -40,8 +47,14 @@ __all__ = [
     "GenerationResponse",
     "BatchGenerationRequest",
     "BatchGenerationResponse",
-    # Services
-    "ContentProcessingService",
+    # Content processing functions
+    "get_content_from_quiz",
+    "prepare_content_for_generation",
+    "prepare_and_validate_content",
+    "validate_content_quality",
+    "validate_module_content",
+    "get_content_statistics",
+    # Generation service
     "QuestionGenerationService",
     "service",
     # Infrastructure