--- a/backend/src/question/services/__init__.py
+++ b/backend/src/question/services/__init__.py
@@ -1,9 +1,9 @@
 """Services module for question generation system."""
 
 from .content_service import ContentProcessingService
-from .generation_service import GenerationOrchestrationService
+from .generation_service import QuestionGenerationService
 
 __all__ = [
     "ContentProcessingService",
-    "GenerationOrchestrationService",
+    "QuestionGenerationService",
 ]