--- a/backend/src/quiz/router.py
+++ b/backend/src/quiz/router.py
@@ -88,7 +88,7 @@ async def create_new_quiz(
         user_id=str(current_user.id),
         canvas_id=current_user.canvas_id,
         canvas_course_id=quiz_data.canvas_course_id,
-        question_count=quiz_data.question_count,
+        question_count=quiz_data.total_question_count,
         llm_model=quiz_data.llm_model,
     )
 
@@ -114,7 +114,7 @@ async def create_new_quiz(
             user_id=str(current_user.id),
             quiz_id=str(quiz.id),
             canvas_course_id=quiz_data.canvas_course_id,
-            question_count=quiz_data.question_count,
+            question_count=quiz_data.total_question_count,
         )
 
         return quiz