… PLAN.md
… backend/src/question/templates/files/batch_multiple_choice.json
… backend/src/question/templates/files/batch_multiple_choice_no.json
… backend/src/question/templates/files/default_multiple_choice.json
… backend/src/question/templates/files/default_multiple_choice_no.json
… backend/src/question/templates/files/enhanced_mcq.json
… backend/src/question/templates/files/enhanced_mcq_no.json
… backend/src/question/workflows/base.py
… backend/src/question/workflows/mcq_workflow.py
… backend/src/question/workflows/module_batch_workflow.py
… backend/src/question/workflows/registry.py
… backend/src/quiz/models.py
… backend/src/quiz/schemas.py
… backend/src/quiz/service.py
… backend/tests/factories.py
… backend/tests/question/services/test_content_service.py
… backend/tests/question/services/test_generation_service.py
… backend/tests/question/workflows/test_mcq_workflow.py
… backend/tests/question/workflows/test_module_batch_workflow.py
… backend/tests/quiz/test_orchestrator.py
… backend/tests/quiz/test_quiz_service.py
… docs/MODULE_BASED_QUESTION_GENERATION.md
… docs/module_based_question_memory.md
… frontend/src/client/types.gen.ts
… frontend/src/components/QuizCreation/ModuleQuestionSelectionStep.tsx
… frontend/src/components/QuizCreation/QuizSettingsStep.tsx
… frontend/src/routes/_layout/create-quiz.tsx
… frontend/src/routes/_layout/quiz.$id.tsx
… frontend/tests/components/module-selection.spec.ts
… frontend/tests/components/quiz-detail-content-extraction.spec.ts